+++ subsys/net/ip/ipv6.h
@@ -385,8 +385,6 @@ typedef void (*net_ipv6_frag_cb_t)(struct net_ipv6_reassembly *reass,
  */
 void net_ipv6_frag_foreach(net_ipv6_frag_cb_t cb, void *user_data);
 
-#endif /* CONFIG_NET_IPV6_FRAGMENT */
-
 /**
  * @brief Find the last IPv6 extension header in the network packet.
  *
@@ -400,6 +398,26 @@ void net_ipv6_frag_foreach(net_ipv6_frag_cb_t cb, void *user_data);
  */
 int net_ipv6_find_last_ext_hdr(struct net_pkt *pkt, u16_t *next_hdr_idx,
 			       u16_t *last_hdr_idx);
+
+/**
+ * @brief Handles IPv6 fragmented packets.
+ *
+ * @param pkt Network head packet.
+ * @param frag Network packet fragment
+ * @param total_len Total length of the packet
+ * @param offset Start of fragment header
+ * @param loc End of fragment header, this is returned to the caller
+ * @param nexthdr IPv6 next header after fragment header part
+ *
+ * @return Return verdict about the packet
+ */
+enum net_verdict net_ipv6_handle_fragment_hdr(struct net_pkt *pkt,
+					      struct net_buf *frag,
+					      int total_len,
+					      u16_t buf_offset,
+					      u16_t *loc,
+					      u8_t nexthdr);
+#endif /* CONFIG_NET_IPV6_FRAGMENT */
 
 #if defined(CONFIG_NET_IPV6)
 void net_ipv6_init(void);
